moved all to the same folder
