+++ src/components/index.ts
@@ -10,5 +10,5 @@ export * from "./Text";
 export * from "./Video";
 export * from "./Fog";
 export * from "./Shopify";
-export * from "./EditMode";
+export * from "./BuilderTools";
 export * from "./TextInput";
